Remove logic to try and hide the comments link from the Meta widget.
`is_active_widget` only tells us whether the meta widget has been added to a sidrbar, not whether that sidebar is actually displayed anywhere. As a result we are injecting useless JS into every page on sites that never need it.

The Meta widget does not provide any way to filter its output, so for now administrators will have to decide themselves whether to use the widget.

Closes #44.